+++ quint/src/effects/inferrer.ts
@@ -18,15 +18,15 @@ import { LookupTable, LookupTableByModule, lookupValue, newTable } from '../look
 import { expressionToString } from '../IRprinting'
 import { IRVisitor, walkModule } from '../IRVisitor'
 import { QuintApp, QuintBool, QuintEx, QuintInt, QuintLambda, QuintLet, QuintModule, QuintModuleDef, QuintName, QuintOpDef, QuintStr } from '../quintIr'
-import { Effect, Name, Signature, effectNames, unify } from './base'
-import { Substitutions, applySubstitution, compose } from './substitutions'
+import { Effect, EffectScheme, Signature, effectNames, toScheme, unify } from './base'
+import { Substitutions, applySubstitution, applySubstitutionToScheme, compose } from './substitutions'
 import { Error, ErrorTree, buildErrorLeaf, buildErrorTree, errorTreeToString } from '../errorTree'
 import { ScopeTree, treeFromModule } from '../scoping'
-import isEqual from 'lodash.isequal'
 import { getSignatures } from './builtinSignatures'
 import { FreshVarGenerator } from '../FreshVarGenerator'
+import { effectToString } from './printing'
 
-export type EffectInferenceResult = [Map<bigint, ErrorTree>, Map<bigint, Effect>]
+export type EffectInferenceResult = [Map<bigint, ErrorTree>, Map<bigint, EffectScheme>]
 
 /* Walks the IR from node to root inferring effects for expressions and
  * assigning them to the effects attribute map, to be used in upward
@@ -53,7 +53,7 @@ export class EffectInferrer implements IRVisitor {
   }
 
   // Public values with results by expression ID
-  private effects: Map<bigint, Effect> = new Map<bigint, Effect>()
+  private effects: Map<bigint, EffectScheme> = new Map<bigint, EffectScheme>()
   private errors: Map<bigint, ErrorTree> = new Map<bigint, ErrorTree>()
 
   private substitutions: Substitutions = []
@@ -106,9 +106,9 @@ export class EffectInferrer implements IRVisitor {
          * ------------------------------------ (NAME-PARAM)
          *          Γ ⊢ v: Read[r_p]
          */
-        let result: Either<Error, Effect>
+        let result: Either<Error, EffectScheme>
         if (def.reference) {
-          result = right({ kind: 'quantified', name: `e_${expr.name}_${def.reference}` })
+          result = right(toScheme({ kind: 'quantified', name: `e_${expr.name}_${def.reference}` }))
         } else {
           result = left(buildErrorLeaf(
             this.location,
@@ -125,7 +125,7 @@ export class EffectInferrer implements IRVisitor {
          *       Γ ⊢ c: Pure
          */
         const effect: Effect = { kind: 'concrete', components: [] }
-        this.addToResults(expr.id, right(effect))
+        this.addToResults(expr.id, right(toScheme(effect)))
         break
       }
       case 'var': {
@@ -136,7 +136,7 @@ export class EffectInferrer implements IRVisitor {
         const effect: Effect = {
           kind: 'concrete', components: [{ kind: 'read', variables: { kind: 'concrete', vars: [{ name: expr.name, reference: expr.id }] } }],
         }
-        this.addToResults(expr.id, right(effect))
+        this.addToResults(expr.id, right(toScheme(effect)))
         break
       }
       case 'val':
@@ -149,10 +149,12 @@ export class EffectInferrer implements IRVisitor {
         if (def.reference) {
           this.addToResults(expr.id, this.fetchResult(def.reference))
         }
-        const result = this.fetchSignature(expr.name, expr.id, 2)
-          .mapLeft(err => buildErrorTree(this.location, err))
 
-        this.addToResults(expr.id, result)
+        this.fetchSignature(expr.name, expr.id, 2)
+          .mapLeft(err => buildErrorTree(this.location, err))
+          .map(effect => {
+            return this.addToResults(expr.id, right(toScheme(effect)))
+          })
       }
     }
   }
@@ -169,9 +171,8 @@ export class EffectInferrer implements IRVisitor {
     }
 
     this.location = `Trying to infer effect for operator application in ${expressionToString(expr)}`
-
     const paramsResult = mergeInMany(expr.args.map((a: QuintEx) => {
-      return this.fetchResult(a.id)
+      return this.fetchResult(a.id).map(e => this.newInstance(e))
     }))
 
     const resultEffect: Effect = { kind: 'quantified', name: this.freshVarGenerator.freshVar('e') }
@@ -185,7 +186,7 @@ export class EffectInferrer implements IRVisitor {
       return effect
     })
 
-    const result = this.fetchSignature(expr.opcode, expr.id, expr.args.length)
+    this.fetchSignature(expr.opcode, expr.id, expr.args.length)
       .mapLeft(err => buildErrorTree(this.location, err))
       .chain(signature => {
         const substitution = arrowEffect.chain(effect => unify(signature, effect))
@@ -196,7 +197,7 @@ export class EffectInferrer implements IRVisitor {
             this.substitutions = s
 
             this.effects.forEach((effect, id) => {
-              const r = applySubstitution(s, effect)
+              const r = applySubstitutionToScheme(s, effect)
               this.addToResults(id, r)
             })
 
@@ -204,16 +205,19 @@ export class EffectInferrer implements IRVisitor {
           })
 
         return resultEffectWithSubs
+      }).map(effect => {
+        this.addToResults(expr.id, right(toScheme(effect)))
+      }).mapLeft(err => {
+        this.addToResults(expr.id, left(err))
       })
 
-    this.addToResults(expr.id, result)
   }
 
   // Literals are always Pure
   exitLiteral(expr: QuintBool | QuintInt | QuintStr): void {
-    this.addToResults(expr.id, right({
+    this.addToResults(expr.id, right(toScheme({
       kind: 'concrete', components: [],
-    }))
+    })))
   }
 
   /*                        Γ ⊢ e: E
@@ -225,10 +229,10 @@ export class EffectInferrer implements IRVisitor {
       // Don't try to infer let if there are errors with the defined expression
       return
     }
-    const e = this.fetchResult(def.expr.id)
+    const result = this.fetchResult(def.expr.id)
 
     // Set the expression effect as the definition effect for it to be available at the result
-    this.addToResults(def.id, e)
+    this.addToResults(def.id, result)
   }
 
   /*     Γ ⊢ e: E
@@ -249,32 +253,52 @@ export class EffectInferrer implements IRVisitor {
    * ---------------------------------------------- (LAMBDA)
    * Γ ⊢ (p0, ..., pn) => e: (E0, ..., En) => E
    */
-  exitLambda(e: QuintLambda): void {
+  exitLambda(lambda: QuintLambda): void {
     if (this.errors.size > 0) {
       return
     }
-    const exprResult = this.fetchResult(e.expr.id)
-    const params = mergeInMany(e.params.map(p => {
-      return this.fetchSignature(p, e.expr.id, 2)
+    const exprResult = this.fetchResult(lambda.expr.id)
+    const params = mergeInMany(lambda.params.map(p => {
+      return this.fetchSignature(p, lambda.expr.id, 2)
         .chain(e => applySubstitution(this.substitutions, e))
     }))
 
-    const result = exprResult.chain(resultEffect => {
-      return params.map((ps): Effect => {
-        return { kind: 'arrow', params: ps, result: resultEffect }
+    exprResult
+      .chain(resultEffect => {
+        return params.map((ps): EffectScheme => {
+          return { ...resultEffect, effect: { kind: 'arrow', params: ps, result: resultEffect.effect } }
+        })
       })
-    })
+      .map(this.newInstance)
+      .chain(effect => applySubstitution(this.substitutions, effect))
+      .map(effect => {
+        if (effect.kind !== 'arrow') {
+          // Impossible
+          throw new Error(`Arrow effect after substitution should be an arrow: ${effectToString(effect)}`)
+        }
 
-    this.addToResults(e.id, result)
+        const nonFreeNames = effect.params.reduce((names, p) => {
+          const { effectVariables, variables } = effectNames(p)
+          return {
+            effectVariables: new Set([...names.effectVariables, ...effectVariables]),
+            variables: new Set([...names.variables, ...variables]),
+          }
+        }, { effectVariables: new Set<string>(), variables: new Set<string>() })
+
+        this.addToResults(lambda.id, right({ ...nonFreeNames, effect }))
+      })
+      .mapLeft(err => {
+        this.addToResults(lambda.id, left(err))
+      })
   }
 
-  private addToResults(exprId: bigint, result: Either<Error, Effect>) {
+  private addToResults(exprId: bigint, result: Either<Error, EffectScheme>) {
     result
       .mapLeft(err => this.errors.set(exprId, buildErrorTree(this.location, err)))
       .map(r => this.effects.set(exprId, r))
   }
 
-  private fetchResult(id: bigint): Either<ErrorTree, Effect> {
+  private fetchResult(id: bigint): Either<ErrorTree, EffectScheme> {
     const successfulResult = this.effects.get(id)
     const failedResult = this.errors.get(id)
     if (failedResult) {
@@ -286,18 +310,16 @@ export class EffectInferrer implements IRVisitor {
     }
   }
 
-
   private fetchSignature(opcode: string, scope: bigint, arity: number): Either<ErrorTree, Effect> {
     // Assumes a valid number of arguments
     if (opcode === '_') {
       return right({ kind: 'quantified', name: this.freshVarGenerator.freshVar('_e') })
     }
 
-    let effect
     if (this.builtinSignatures.has(opcode)) {
       const signatureFunction = this.builtinSignatures.get(opcode)!
-      effect = signatureFunction(arity)
-      return right(this.newInstance(effect))
+      const signature = signatureFunction(arity)
+      return right(this.newInstance(signature))
     } else {
       const def = lookupValue(this.currentTable, this.currentScopeTree, opcode, scope)
       const id = def?.reference
@@ -310,28 +332,21 @@ export class EffectInferrer implements IRVisitor {
       }
 
       return this.fetchResult(id).map(e => {
-        if (e.kind === 'arrow') {
-          return this.newInstance(e)
-        } else {
-          return e
-        }
+        return this.newInstance(e)
       })
     }
   }
 
-  private newInstance(effect: Effect): Effect {
-    const names: Name[] = effectNames(effect)
-    const uniqueNames: Name[] = []
-    names.forEach(name => {
-      if (!uniqueNames.some(n => isEqual(n, name))) {
-        uniqueNames.push(name)
-      }
+  private newInstance(effect: EffectScheme): Effect {
+    const effectSubs: Substitutions = [...effect.effectVariables].map(name => {
+      return { kind: 'effect', name: name, value: { kind: 'quantified', name: this.freshVarGenerator.freshVar('e') } }
     })
-    const subs: Substitutions = uniqueNames.map(name => {
-      return { kind: name.kind, name: name.name, value: { kind: 'quantified', name: this.freshVarGenerator.freshVar('v') } }
+    const variableSubs: Substitutions = [...effect.variables].map(name => {
+      return { kind: 'variable', name: name, value: { kind: 'quantified', name: this.freshVarGenerator.freshVar('v') } }
     })
 
-    const result = applySubstitution(subs, effect)
+    const result = compose(effectSubs, variableSubs).chain(s => applySubstitution(s, effect.effect))
+
     if (result.isLeft()) {
       throw new Error(`Error applying fresh names substitution: ${errorTreeToString(result.value)} `)
     } else {